chore: Rename Task to Assignment
