+++ data/assignment/src/main/java/com/seugi/data/assignment/di/RepositoryModule.kt
@@ -1,0 +1,18 @@
+package com.seugi.data.assignment.di
+
+import com.seugi.data.assignment.AssignmentRepository
+import com.seugi.data.assignment.repository.AssignmentRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface RepositoryModule {
+
+    @Singleton
+    @Binds
+    fun bindsAssignmentRepository(taskRepositoryImpl: AssignmentRepositoryImpl): AssignmentRepository
+}
