+++ data/assignment/src/main/java/com/seugi/data/assignment/model/AssignmentType.kt
@@ -1,0 +1,6 @@
+package com.seugi.data.assignment.model
+
+enum class AssignmentType {
+    GOOGLE,
+    WORKSPACE,
+}
