--- data/task/src/main/java/com/seugi/data/task/model/TaskType.kt
@@ -1,6 +1,0 @@
-package com.seugi.data.task.model
-
-enum class TaskType {
-    GOOGLE,
-    WORKSPACE,
-}
