+++ feature-main/assignment/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    alias(libs.plugins.seugi.android.feature)
+    alias(libs.plugins.seugi.android.hilt)
+    alias(libs.plugins.seugi.android.kotlin)
+}
+
+android {
+    namespace = "com.seugi.task"
+}
+dependencies {
+    implementation(libs.kotlinx.collections.immutable)
+    implementation(libs.kotlinx.datetime)
+
+    implementation(projects.designsystem)
+    implementation(projects.common)
+    implementation(projects.ui)
+    implementation(projects.data.assignment)
+}
