+++ feature-main/main/src/main/java/com/seugi/main/MainScreen.kt
@@ -46,8 +46,8 @@ import com.seugi.setting.navigate.navigateToSetting
 import com.seugi.setting.navigate.settingScreen
 import com.seugi.task.create.navigation.navigateToTaskCreate
 import com.seugi.task.create.navigation.taskCreateScreen
-import com.seugi.task.navigation.navigateToTask
-import com.seugi.task.navigation.taskScreen
+import com.seugi.assignment.navigation.navigateToAssignment
+import com.seugi.assignment.navigation.assignmentScreen
 import com.seugi.timetable.navigation.navigateToTimetable
 import com.seugi.timetable.navigation.timetableScreen
 import com.seugi.workspace.navigation.WAITING_JOIN
@@ -157,7 +157,7 @@ internal fun MainScreen(
                     navHostController.navigateToWorkspaceCreate()
                 },
                 navigateToTask = {
-                    navHostController.navigateToTask()
+                    navHostController.navigateToAssignment()
                 },
                 navigateToMeal = navHostController::navigateToMeal,
             )
@@ -362,7 +362,7 @@ internal fun MainScreen(
                 workspaceId = state.profile.workspaceId,
             )
 
-            taskScreen(
+            assignmentScreen(
                 popBackStack = navHostController::popBackStack,
                 workspaceId = state.workspace.workspaceId,
                 profile = state.profile,
