+++ network/assignment/src/main/java/com/seugi/network/assignment/response/AssignmentGoogleResponse.kt
@@ -1,0 +1,11 @@
+package com.seugi.network.assignment.response
+
+import java.time.LocalDateTime
+
+data class AssignmentGoogleResponse(
+    val id: Long,
+    val title: String,
+    val description: String?,
+    val link: String,
+    val dueDate: LocalDateTime?,
+)
