+++ network/assignment/src/main/java/com/seugi/network/assignment/response/AssignmentResponse.kt
@@ -1,0 +1,11 @@
+package com.seugi.network.assignment.response
+
+import java.time.LocalDateTime
+
+data class AssignmentResponse(
+    val id: Long,
+    val workspaceId: String,
+    val title: String,
+    val description: String?,
+    val dueDate: LocalDateTime?,
+)
