--- network/task/src/main/java/com/seugi/network/task/request/TaskCreateRequest.kt
@@ -1,10 +1,0 @@
-package com.seugi.network.task.request
-
-import java.time.LocalDateTime
-
-data class TaskCreateRequest(
-    val workspaceId: String,
-    val title: String,
-    val description: String,
-    val dueDate: LocalDateTime,
-)
