--- network/task/src/main/java/com/seugi/network/task/response/TaskResponse.kt
@@ -1,11 +1,0 @@
-package com.seugi.network.task.response
-
-import java.time.LocalDateTime
-
-data class TaskResponse(
-    val id: Long,
-    val workspaceId: String,
-    val title: String,
-    val description: String?,
-    val dueDate: LocalDateTime?,
-)
